fix(toolkit): multi-stack apps cannot be synthesized or deployed (#911)
Due to a recent cx protocol change (#868), some toolkit commands stopped
respecting the "selected" stacks (the ones specified in the command line).

"cdk synth" would always return the first stack, and "cdk deploy" would
always deploy all stacks.

Since we have test coverage gaps in the toolkit (#294), we did not
discover this before we released.

This change includes an initial set of integration tests for the 
toolkit. At the moment they should be manually executed when toolkit
changes are made, but we will execute them in a pipeline.

Fixes #910