chore: make toolkit tests not break Lerna links
If the integ tests are run in the source directroy, the "npm install"
command will go up to find package.json and replace its node_modules
directory, thereby breaking the Lerna symlinks and making subsequent
toolkit executions fail because it can't find '@aws-cdk/cx-api'.

Copy out the app directory to a temporary directory and run the test
there, so that the source tree remains undisturbed.

It's a fixed directory name, seems simpler than generating a temporary
directory on each run and then also having to install a cleanup hook.