fix: Auto-delete stacks that failed creating before new attempt
Restore the previous behavior around stacks that failed creation,
detecting the condition and automatically deleting the stack before
making a new attempt.

The stack is not deleted after the failure in order to allow for
forensics to be performed. Also, deleting the stack on the next creation
attempt is more robust.