feat(toolkit): Produce an Asset manifest document
Consumes the metadata produced by `Asset`s into the `SynthesizedStack`s,
and builds an asset manifest document and "internalizes" the supporting files.
The supporting files are de-duplicated using a `sha-256`, so that the same
file(s) referenced from multiple assets or stacks will be presented only once in
the output.

The asset document maps an asset logical ID to the "internalized" file(s) base
path in the output directory, the construct path where the metadata was created,
and the output parameters specified by the asset (currently always an S3
bucket and key).