feat(core): Pre-concatenate Fn::Join
This restores the pre-concatenation of `Fn::Join` fragments that was
introduced in #916 and reverted in #958 due to an other feature part of
the same commit.

The goal is to pre-concatenate the string literals present in the
concatenated values, and flatten `Fn::Join` nested under another
`Fn::Join` using the same delimiter, so as to produce more readable
CloudFormation templates. Additionally, if the `Fn::Join`
pre-concatenation results in a single string being left, then the
`Fn::Join` intrinsic is erased and replaced with a string literal.