Add instance data to use in node group creation
This file maps AMI based on instance chosen, normal or gpu.
It also maps the max pods per node based on AWS cfn template
https://amazon-eks.s3-us-west-2.amazonaws.com/cloudformation/2018-08-30/amazon-eks-nodegroup.yaml

I tried using a context search, but the account ID for AMIs is different
The path also seems to be different, will need to investigate.
This shouldn't be a problem for now, as it's only a total of 6 AMIs