feat(aws-sqs): allow direct grants from a Queue
Add grantReceive and grantSend to a Queue allowing a user to be able to
directly grant permissions to an IAM Principal (Role, User etc.).

An additional grantConsume is here for combining three permissions
commonly used when consuming messages rather than just receiving to
provide the ability to set message visibility and delete entirely.