fix(aws-s3-deployment): avoid deletion during update using physical ids (#1006)
When a BucketDeployment resource is created, a unique physical ID is
generated and returned via `ResourcePhysicalId`. The same ID will then
be used in update/delete.

This tells CloudFormation not to issue a DELETE operation after an UPDATE,
which was the cause for #981.

Also, allow destination prefix to be "/", which is
the same as not specifying a prefix.

Fixes #981