feat(cicd): CI/CD for CDK Stacks
This is a very simple construct that allows one to add actions to a
CodePipeline to deploy a CDK stack.

Currently comes with some pretty strong limitations (deployed stack &
code pipeline must reside in the same account & region, assets are not
supported).

It does however lay the foundation for the API, as the external shape of
the feature shouldn't change (or not much) when amending to support
cross-account deployment and assets.