feat: add a new construct library for ECS (#1058)
Add a new construct library to start services on ECS, both on EC2 and
Fargate-based clusters.

Containers can be started from images that are publicly available on
DockerHub, images in ECR repositories, and images built directly from
sources that are stored in source code next to the CDK app.

ECS services can be used as load balancer targets, and there are
higher-level constructs available to make it easy to start a service
behind a load balancer.

BREAKING CHANGE: the ec2.Connections object has been changed to be able
to manage multiple security groups. The relevant property has been
changed from `securityGroup` to `securityGroups` (an array of security
group objects).