fix(aws-ecs): fix use of published NPM package with TypeScript (#1117)
The package previously published would not work with a TypeScript consumer, because the generated `.d.ts` file was missing an import statement due to a bug in the TypeScript compiler. Add an explicit type declaration to work around this issue until it is fixed and released upstream.

Reference: https://github.com/Microsoft/TypeScript/issues/26969