feat(cdk): allow Tokens to be encoded as lists (#1144)
Add the ability for Tokens to be encoded into lists using the
`token.toList()` method.

This is useful for Tokens that intrinsically represent lists of strings,
so we can pass them around in the type system as `string[]` (and
interchange them with literal `string[]` values).

The encoding is reversible just like string encodings are reversible.
Contrary to strings encodings, concatenation operations are not
allowed (they cannot be applied after decoding since CloudFormation
does not have any list concatenation operators).

This change does not change any CloudFormation resources yet to
take advantage of the new encoding.

Implements the most important remaining part of #744.