chore: fix Python build of s3-deployent on macOS (#1141)
On macOS developers are likely to use a Homebrew Python,
which comes with a default pip installation prefix configured.

On those systems, running 'pip install --target' leads to the following
error:

DistutilsOptionError: must supply either home or prefix/exec-prefix — not both

Override the system prefix to be empty, and just force
pip to install in the given directory.