fix(aws-ec2): add dependency on gateway attachment for public routes (#1142)
Currently there is a race condition when creating the
public subnet routes for the ec2.VpcNetwork construct. CloudFormation
can attempt to create the public subnet routes to the IGW before the IGW
is attached to the VPC. This change adds a dependency to the public
routes on the IGW attachment.

Fixes #1140. 

BREAKING CHANGE: Method signature of
VpcPublicSubnet.addDefaultIGWRouteEntry changed in order to add a
dependency on gateway attachment completing before creating the public
route to the gateway. Instead of passing a gateway ID string, pass in a
cloudformation.InternetGatewayResource object and a
cloudformation.VPCGatewayAttachmentResource object.