fix(aws-cdk): make bootstrapping not require --app
Improve the error message when needing to bootstrap while using assets.
Also make it not necessary to have '--app' in there when bootstrapping
if an environment name is given.

This makes the toolkit easier to use in integration tests, where the
cdk.json is generated and it's not clear what the argument to --app
should be (nor should it be necessary if the environment identifier has
enough information in it to reconstruct the Environment object).

This also refactors the code somewhat so related code (for running the
CX App and parsing its output) is together.

Fixes #1188.