chore: Add rule to guarantee node runtime compatibility (#1205)
This adds a rule to make sure the codebase is all configured to support
node runtimes complying with `>= 8.11.0`, by ensuring that the
configured `@types/node` version starts with `^8.`.

In order to allow applying this rule to the top-level `package.json`, a
feature was added to `pkglint` that allows users to confgure `include`
and `exclude` filters as part of their `package.json`'s `pkglint` setup.
Exclusions always have precedence over inclusions, and specifying
`disable: true` has the same effect as specifying `exclude: '*'`. Both
the `include` and `exclude` values can contain patterns that use `*` as
a wild-card that maches any number of characters (including no
character at all), and can be either a single pattern, or an array of
patterns. `include` defaults to `*`.

Fixes #1194