chore: refactoring of toolkit and cxapi
Right now we're abusing the 'command objects' feature of yargs with a
decorator that changes the expected type of the handler function from
(yargs) => void to (yargs) => Promise<number>.

Instead, we make explicit what is happening: during the execution of the
original handler, we store the async "real" handler object is written to
a variable and executed later on instead of main().

The async handler gets access to more of the state of the application so
that more realistic handlers can be written that need access to the CX
app's stacks.

Organize metadata types of CX API into a folder.