Make loading of specification.json lazy
Before, the json file would be loaded upon initialization of the
library, but the library would also be loaded in the build tool.

It's only because of some crazy lazy evaluation optimization done
by NodeJS that that didn't blow up before, but due to code changes it
does blow up now.