fix(aws-elbv2): add dependency on internet gateway
Internet-facing load balancers can only be created once the VPC's
Internet Gateway has been created, otherwise creation is an error.

Usually the gateway is speedy enough to create so that this lack of a
dependency doesn't show, but to be correct we still need to add it.

Fixes #1129.

ALSO IN THIS CHANGE

- Add a script to visualize dependencies in a CloudFormation template
  (by converting to .dot).