fix(assert): Correct bug in SUPERSET matching
The SUPERSET matching mode would have aborted successfully upon finding
the first acceptable change (a change that is a CREATE), and not look at
subsequent changes, hence allowing any other type of update to pass as
long as it is detected later in the template.

This change fixes that such that the assertion will reject upon any
change that is NOT a CREATE change.

Fixes awslabs/cdk-ops#186