better reading
AWS Construct Library members are found in the @aws-cdk/NAMESPACE packages, where NAMESPACE is the short name for the associated service, such as SQS for the AWS Construct Library for the Amazon SQS service. 
->
Members of the AWS Construct Library are found in the @aws-cdk/NAMESPACE packages, where NAMESPACE is the short form of the associated service, such as SQS for the Amazon SQS service.