fix(toolkit): scrutiny dialog should fail with no tty (#1382)
If STDIN is not connected to a TTY (terminal), and scrutiny is enabled,
we expect the program to fail (exit with non-zero exit code).

This is especially important for CI/CD scenarios where you wouldn't want
to accidentally deploy changes that didn't pass a scrutiny check.

Added integration test.

Fixes #1380