feat(cli): Allow specifying options using env vars (#1447)
Allows passing options to the `cdk` command using environment variables
with the `CDK_` name prefix. This can be used, for example, for the
following:

- `CDK_VERSION_REPORTING=false` disables version reporting
- `CDK_REQUIRE_APPROVAL=broadening` requires approval for changes to
  security policies that broaden the grants
- `CDK_VERBOSE=true` enables verbose logging in the toolkit

Added a validator in `cdk doctor` that will display the environment
variables that may be picked up by the toolkit, and highlight those that
are reserved for internal use by the toolkit (`CDK_CONTEXT_JSON` and
`CDK_OUTDIR`).