feat(diff): Better diff of random objects (#1488)
Use a unified diff format to render differences in arbitrary values,
making it easier to understand what is changing in possibly large JSON
structures, for example.

The number of context lines used when rendering the JSON differences
can be customized using the `--context-lines` option of `cdk diff`, 
which has a default value of `3`.