Change replacement diffing to not trigger IAM changes dialog
We used to modify the new template in place, causing changes to
"logicalIDs" of replaced resources, which would trigger downstream
changes and potential downstream replacements, etc.

This would lead to the changed logical ID popping up in the IAM
diff which was not desirable.

In this change, we do the same processing but we throw away the
template after all changes have been propagated, and only copy
the resultant property STATUSES onto the diff object that gets
returned to the user. This leads to the same displayed result
without the changes to the template actually propagating.

In the course of this modification, the diff classes have been changed
to also have objects in places of resources and properties that
didn't actually change (so that they could be modified in-place),
and diff objects have a boolean telling whether they are actual
changes or not.

Fixes #1495.