rfc: external resource references
now that we have automatic references of resources across stacks within
the same app, we are revisiting our model for referencing external resources.

this rfc proposes to add a serialization scheme for cdk constructs
and on top of it implement cross-app export/imports. it further describes
how external resources can be referenced via a physical resource name
such as an arn or a name, and proposes naming conventions for the
various elements.