feat(core): allow multiple Apps
Change the meaning of `App` to represent the application the user
is defining (as opposed to the internals of the CDK program). The
names of Apps (if given) will be incorporated into the Stack name,
so that it becomes possible to instantiate your Stacks multiple
times just like other constructs (without having to make the
`id` parameter unique). A new stack property `stackName` exists
to force the stack to a different name.

If a single `App` is instantiated without parameters, the observable
behavior of your CDK app won't change. If you insantiate more than
one `App`, all of them must receive ids.

To represent the CDK internals, a new class called `Program` exists,
which the user doesn't need to interact with.

`app.run()` is no longer necessary, and is deprecated. It still
exists to avoid breaking existing programs, but will be removed
in a future release.