feat(aws-codepipeline): support setting a Role for a CFN Action (#1449)
This adds the option to set a Role for a CodePipeline Action.
As of now, this is limited to only the CloudFormation Pipeline Actions.
Setting the Role is needed for some more advanced setups,
for example cross-account deployments.

BREAKING CHANGE:  the `role` property in the CloudFormation Actions has been renamed to `deploymentRole`.
BREAKING CHANGE: the `role` property in the `app-delivery` package has been renamed to `deploymentRole`.