self-contained
turn decdk to a self-contained transformer which depends on all cdk
libraries to simplify usage.

theoretically we would have been able to split decdk and the dependency
closure, but keeping them monolithic right now is easier to work with
and takes care of some issues related to resolving modules
in client packages (e.g. we need to resolve transitive modules and
that's annoying).