support Fn::GetAtt for resources and constructs
if we see an Fn::GetAtt, we replace it with a synth-time lazy value
which will resolve according to what is referenced.

if GetAtt references a high level construct, the resolved value will
be the value returned by a named property. otherwise, we just retain
the original Fn::GetAtt.

at the moment, this is only supported for strings.