apigateway: resourceForPath
a new "resourceForPath(path)" method will get or create all
the resources leading to a certain path in the api model.

this can be used to easily define "routes" in the api and
reuse the nodes in the tree.

includes a refactor of how the root resource and normal resources
share implementation (through an abstract base class, dah!)