deconstructing enum-like classes
Classes that have static methods/properties that return
a type that matches the class type can be considered "enum-like"
because they can be used like so:

    lambda.Runtime.NodeJS10

or:

    lambda.Code.asset('path')

We are now deconstructing these to either a string (for the property
case) or an object where the key is the method name and the value
is an map of parameter names to values.

This is actually something we can probably embed in the jsii spec,
as it is a common pattern.