ecs: fix ctor signature of ContainerDefinition
in order to be able to deconstruct this, we expect ctors to have only three arguments: scope,id,props.

split ContainerDefinitionOptions from ContainerDefinitionProps to allow
reusing them in `addContainer`.

also, move the logic of "linking" the container to the task definition
into the ContainerDefinition class, so that "addContainer" is pure
sugar and doesn't have special behavior. otherwise, we won't be able
to use it declaratively.